src/code/requirements.txt
pillow==9.3
boto3
